Create MAPAMA implementation  (EIDOS system)
An implementation of Plinian Core as the official repository for species information in Spain: the EIDOS system at the Direction General of Biodiversity and Environmental Quality (Ministry of Ecological Transition of Spain). The following link does not contain a xsd file, but documentation on the database developed to support Plinian Core elements and to be compliant with it.